+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- Added add ItemSearch.url_with_parameters to get a formatted search url` [#259](https://github.com/stac-utils/pystac-client/issues/299)
+
 ### Changed
 
 ### Fixed